requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attr==0.3.2
attrs==25.1.0
bcrypt==4.0.1
cffi==1.17.1
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.1.1
idna==3.10
makefun==1.15.6
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
pwdlib==0.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
